requirements.txt
Flask==2.0.1
Flask_Bcrypt==1.0.1
flask_limiter==4.1.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
pandas==3.0.1
Pillow==12.1.1
Werkzeug==2.0.1
gunicorn
